+++ src/QueryCat.Backend/Formatters/FormattersInfo.cs
@@ -9,47 +9,28 @@ namespace QueryCat.Backend.Formatters;
 /// </summary>
 public static class FormattersInfo
 {
-    private static readonly Dictionary<string, Func<ExecutionThread, IRowsFormatter>> Formatters = new()
+    private static readonly Dictionary<string, Func<FunctionsManager, FunctionArguments, IRowsFormatter>> Formatters = new()
     {
         // File extensions.
-        [".csv"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
-        [".tsv"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
-                FunctionArguments.Create().Add("delimiter", '\t')),
-        [".tab"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
-                FunctionArguments.Create().Add("delimiter", '\t')),
-        [".log"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
-                FunctionArguments.Create().Add("delimiter", ' ').Add("delimiter_can_repeat", false)),
-        [".json"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("json"),
-        [".xml"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
-        [".xsd"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
+        [".csv"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args),
+        [".tsv"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args.Add("delimiter", '\t')),
+        [".tab"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args.Add("delimiter", '\t')),
+        [".log"] = (fm, args) =>
+            fm.CallFunction<IRowsFormatter>("csv", args.Add("delimiter", ' ').Add("delimiter_can_repeat", false)),
+        [".json"] = (fm, args) => fm.CallFunction<IRowsFormatter>("json", args),
+        [".xml"] = (fm, args) => fm.CallFunction<IRowsFormatter>("xml", args),
+        [".xsd"] = (fm, args) => fm.CallFunction<IRowsFormatter>("xml", args),
 
         // Content types.
-        ["text/csv"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
-        ["text/x-csv"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
-        ["application/csv"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
-        ["application/x-csv"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
-        ["text/tab-separated-values"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
-                FunctionArguments.Create().Add("delimiter", '\t')),
-        ["application/json"] = thread=>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("json"),
-        ["application/xml"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
-        ["application/xhtml+xml"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
-        ["application/soap+xml"] = thread =>
-            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
+        ["text/csv"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args),
+        ["text/x-csv"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args),
+        ["application/csv"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args),
+        ["application/x-csv"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args),
+        ["text/tab-separated-values"] = (fm, args) => fm.CallFunction<IRowsFormatter>("csv", args.Add("delimiter", '\t')),
+        ["application/json"] = (fm, args)=> fm.CallFunction<IRowsFormatter>("json", args),
+        ["application/xml"] = (fm, args) => fm.CallFunction<IRowsFormatter>("xml", args),
+        ["application/xhtml+xml"] = (fm, args) => fm.CallFunction<IRowsFormatter>("xml", args),
+        ["application/soap+xml"] = (fm, args) => fm.CallFunction<IRowsFormatter>("xml", args),
     };
 
     /// <summary>
@@ -57,12 +38,13 @@ public static class FormattersInfo
     /// </summary>
     /// <param name="id">File extension or content type.</param>
     /// <param name="thread">Execution thread.</param>
+    /// <param name="args">Arguments.</param>
     /// <returns>Instance of <see cref="IRowsFormatter" /> or null.</returns>
-    public static IRowsFormatter? CreateFormatter(string id, ExecutionThread thread)
+    public static IRowsFormatter? CreateFormatter(string id, ExecutionThread thread, FunctionArguments? args = null)
     {
         if (Formatters.TryGetValue(id.ToLower(), out var factory))
         {
-            return factory.Invoke(thread);
+            return factory.Invoke(thread.FunctionsManager, args ?? new FunctionArguments());
         }
         return null;
     }
@@ -72,7 +54,7 @@ public static class FormattersInfo
     /// </summary>
     /// <param name="id">Identifier (file extension or content type).</param>
     /// <param name="formatterFunc">Delegate to create formatter.</param>
-    public static void RegisterFormatter(string id, Func<ExecutionThread, IRowsFormatter> formatterFunc)
+    public static void RegisterFormatter(string id, Func<FunctionsManager, FunctionArguments, IRowsFormatter> formatterFunc)
     {
         Formatters[id] = formatterFunc;
     }