+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add "NoSpaceTable" output style.
 - Support strings unescape.
 - Add more arguments to function "csv".
+- Allow to pass arguments to formatters. Example: "SELECT * FROM '1.csv??has_header=false'".
 
 ### Fixed